backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "PlanTracker app backend"
authors = ["Timur Zheksimbaev <113367706+TimurZheksimbaev@users.noreply.github.com>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
sqlalchemy = "^2.0.40"
pydantic = {extras = ["email"], version = "^2.11.4"}
python-jose = "^3.4.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
flake8 = "^7.2.0"
bcrypt = "^4.3.0"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.1.0"
aiogram = "^3.20.0.post0"
httpx = "^0.28.1"
autopep8 = "^2.3.2"
pytest-asyncio = "^0.26.0"
python-telegram-bot = "^22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

